+++ app/src/main/java/com/example/call_mapbox_api/data/local/EvPointLocalDataSource.kt
@@ -1,6 +1,5 @@
 package com.example.call_mapbox_api.data.local
 
-import com.example.call_mapbox_api.data.remote.EvPointsEntity
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 import javax.inject.Singleton
