+++ app/src/main/java/com/example/call_mapbox_api/data/local/EvPointsDao.kt
@@ -4,7 +4,6 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.example.call_mapbox_api.data.remote.EvPointsEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
