+++ app/src/main/java/com/example/call_mapbox_api/data/local/EvPointsEntity.kt
@@ -1,34 +1,72 @@
-package com.example.call_mapbox_api.model
+package com.example.call_mapbox_api.data.local
 
-import com.example.call_mapbox_api.data.remote.AddressInfo
-import com.example.call_mapbox_api.data.remote.Connections
-import com.example.call_mapbox_api.data.remote.EvPointsEntity
+
+import androidx.annotation.Keep
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import com.example.call_mapbox_api.data.remote.dto.AddressInfo
+import com.example.call_mapbox_api.data.remote.dto.Connections
+import com.example.call_mapbox_api.domain.model.EvPointDetails
 import kotlinx.serialization.Serializable
 
+@Keep
 @Serializable
-data class EvPointDetails(
+@Entity(tableName = "evPointsDb")
+data class EvPointsEntity(
+    @PrimaryKey(autoGenerate = false)
     val ID: Int?,
-    val DateLastVerified: String,
-    val DataProviderID: Int?,
-    val DataQualityLevel: Int?,
-    val DateCreated: String?,
-    val DateLastStatusUpdate: String?,
-    val NumberOfPoints: Int?,
-    val IsRecentlyVerified: Boolean?,
-    val OperatorID: Int?,
-    val StatusTypeID: Int?,
-    val SubmissionStatusTypeID: Int?,
-    val UUID: String?,
-    val UsageCost: String?,
-    val UsageTypeID: Int?,
-    val Connections: List<Connections>,
-    val AddressInfo: AddressInfo
 
+    @ColumnInfo(name = "data_last_verified")
+    val DateLastVerified: String,
+
+    @ColumnInfo(name = "data_provider_id")
+    val DataProviderID: Int?,
+
+    @ColumnInfo(name = "data_quality_level")
+    val DataQualityLevel: Int?,
+
+    @ColumnInfo(name = "data_created")
+    val DateCreated: String?,
+
+    @ColumnInfo(name = "data_last_status_update")
+    val DateLastStatusUpdate: String?,
+
+    @ColumnInfo(name = "number_of_points")
+    val NumberOfPoints: Int?,
+
+    @ColumnInfo(name = "is_recently_verified")
+    val IsRecentlyVerified: Boolean?,
+
+    @ColumnInfo(name = "operator_id")
+    val OperatorID: Int?,
+
+    @ColumnInfo(name = "status_type_id")
+    val StatusTypeID: Int?,
+
+    @ColumnInfo(name = "submission_status_type_id")
+    val SubmissionStatusTypeID: Int?,
+
+    @ColumnInfo(name = "uuid")
+    val UUID: String?,
+
+    @ColumnInfo(name = "usage_cost")
+    val UsageCost: String?,
+
+    @ColumnInfo(name = "usage_type_id")
+    val UsageTypeID: Int?,
+
+    @ColumnInfo(name = "connection")
+    val Connections: List<Connections>,
+
+    @ColumnInfo(name = "address_info")
+    val AddressInfo: AddressInfo
 )
 
-fun List<EvPointDetails>.toEvPointsBreakItems() : List<EvPointsEntity> {
+
+fun List<EvPointsEntity>.toEvPointDetails() : List<EvPointDetails> {
     return this.map {
-        EvPointsEntity(
+        EvPointDetails(
             AddressInfo = it.AddressInfo,
             Connections = it.Connections,
             NumberOfPoints = it.NumberOfPoints,
@@ -48,5 +86,3 @@ fun List<EvPointDetails>.toEvPointsBreakItems() : List<EvPointsEntity> {
         )
     }
 }
-
-
