+++ app/src/main/java/com/example/call_mapbox_api/data/repository/ISearchListRepository.kt
@@ -1,6 +1,6 @@
 package com.example.call_mapbox_api.data.repository
 
-import com.example.call_mapbox_api.data.remote.EvPointsEntity
+import com.example.call_mapbox_api.data.local.EvPointsEntity
 import kotlinx.coroutines.flow.Flow
 
 interface ISearchListRepository {
