+++ app/src/main/java/com/example/call_mapbox_api/domain/ISearchListUseCase.kt
@@ -1,7 +1,6 @@
 package com.example.call_mapbox_api.domain
 
-import com.example.call_mapbox_api.data.remote.EvPointsEntity
-import com.example.call_mapbox_api.model.EvPointDetails
+import com.example.call_mapbox_api.data.local.EvPointsEntity
 import kotlinx.coroutines.flow.Flow
 
 interface ISearchListUseCase {
