+++ app/src/main/java/com/example/call_mapbox_api/domain/SearchListUseCaseImpl.kt
@@ -1,6 +1,6 @@
 package com.example.call_mapbox_api.domain
 
-import com.example.call_mapbox_api.data.remote.EvPointsEntity
+import com.example.call_mapbox_api.data.local.EvPointsEntity
 import com.example.call_mapbox_api.data.repository.ISearchListRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
