+++ app/src/main/java/com/example/call_mapbox_api/domain/model/EvPointDetails.kt
@@ -1,70 +1,34 @@
-package com.example.call_mapbox_api.data.remote
+package com.example.call_mapbox_api.domain.model
 
-
-import androidx.annotation.Keep
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.example.call_mapbox_api.model.EvPointDetails
+import com.example.call_mapbox_api.data.local.EvPointsEntity
+import com.example.call_mapbox_api.data.remote.dto.AddressInfo
+import com.example.call_mapbox_api.data.remote.dto.Connections
 import kotlinx.serialization.Serializable
 
-@Keep
 @Serializable
-@Entity(tableName = "evPointsDb")
-data class EvPointsEntity(
-    @PrimaryKey(autoGenerate = false)
+data class EvPointDetails(
     val ID: Int?,
-
-    @ColumnInfo(name = "data_last_verified")
     val DateLastVerified: String,
-
-    @ColumnInfo(name = "data_provider_id")
     val DataProviderID: Int?,
-
-    @ColumnInfo(name = "data_quality_level")
     val DataQualityLevel: Int?,
-
-    @ColumnInfo(name = "data_created")
     val DateCreated: String?,
-
-    @ColumnInfo(name = "data_last_status_update")
     val DateLastStatusUpdate: String?,
-
-    @ColumnInfo(name = "number_of_points")
     val NumberOfPoints: Int?,
-
-    @ColumnInfo(name = "is_recently_verified")
     val IsRecentlyVerified: Boolean?,
-
-    @ColumnInfo(name = "operator_id")
     val OperatorID: Int?,
-
-    @ColumnInfo(name = "status_type_id")
     val StatusTypeID: Int?,
-
-    @ColumnInfo(name = "submission_status_type_id")
     val SubmissionStatusTypeID: Int?,
-
-    @ColumnInfo(name = "uuid")
     val UUID: String?,
-
-    @ColumnInfo(name = "usage_cost")
     val UsageCost: String?,
-
-    @ColumnInfo(name = "usage_type_id")
     val UsageTypeID: Int?,
-
-    @ColumnInfo(name = "connection")
     val Connections: List<Connections>,
-
-    @ColumnInfo(name = "address_info")
     val AddressInfo: AddressInfo
+
 )
 
-
-fun List<EvPointsEntity>.toEvPointDetails() : List<EvPointDetails> {
+fun List<EvPointDetails>.toEvPointsBreakItems() : List<EvPointsEntity> {
     return this.map {
-        EvPointDetails(
+        EvPointsEntity(
             AddressInfo = it.AddressInfo,
             Connections = it.Connections,
             NumberOfPoints = it.NumberOfPoints,
@@ -84,3 +48,5 @@ fun List<EvPointsEntity>.toEvPointDetails() : List<EvPointDetails> {
         )
     }
 }
+
+
