+++ app/src/main/java/com/example/call_mapbox_api/ui/detailscreen/DetailFragment.kt
@@ -4,15 +4,10 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
-import com.example.call_mapbox_api.R
 import com.example.call_mapbox_api.databinding.FragmentDetailBinding
-import com.example.call_mapbox_api.databinding.FragmentMapbarBinding
-import com.example.call_mapbox_api.databinding.FragmentSearchListBinding
 import com.example.call_mapbox_api.ui.searchscreen.SearchListViewModel
 import com.example.call_mapbox_api.util.SPAN_COUNT
 import dagger.hilt.android.AndroidEntryPoint
